requirements.txt
Flask==1.1.2
line-bot-sdk==1.8.0
pymysql==0.9.2